tup-376-moved-proj-details-to-tup-components
