+++ libs/tup-components/src/projects/details/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectDetails } from './ProjectDetails';
