+++ libs/tup-components/src/projects/index.ts
@@ -3,3 +3,5 @@ export { default as ProjectsDashboard } from './ProjectsDashboard';
 export { default as UserList } from './users/UserList/UserList';
 export { default as UserDetail } from './users/UserDetail/UserDetail';
 export { ProjectsNavbar, ProjectsListing } from './ProjectsListing';
+export { default as ProjectDetails } from './details/ProjectDetails';
+
